+++ plugins/postcss-logical/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added: Support for `block-size`, `inline-size` and `caption-side`.
 - Removed: Support for the `logical` keyword within `margin`, `padding`, `border-width`, `border-style`, `border-color` properties. (major)
 - Removed support for `preserve` option. Changes are no longer preserved (major).
+- Removed support for transforming `float` and `clear` properties. (major)
 - Updated: The plugin now requires block and inline direction to be configured (uses `top-to-bottom` and `left-to-right` respectively) and no longer introduces `:dir` (major).
 
 ### 5.0.4 (February 5, 2022)